+++ nequip/data/dataset.py
@@ -2,14 +2,18 @@ import numpy as np
 import logging
 import tempfile
 import inspect
+import functools
+import itertools
 import yaml
 import hashlib
 from os.path import dirname, basename, abspath
 from typing import Tuple, Dict, Any, List, Callable, Union, Optional, Sequence
 
 import ase
+import ase.io
 
 import torch
+import torch.multiprocessing as mp
 
 from torch_runstats.scatter import scatter_std, scatter_mean
 
@@ -27,6 +31,7 @@ from nequip.data import (
 from nequip.utils.batch_ops import bincount
 from nequip.utils.regressor import solver
 from nequip.utils.savenload import atomic_write
+from nequip.utils.multiprocessing import num_tasks
 from .transforms import TypeMapper
 from .AtomicData import _process_dict
 
@@ -727,6 +732,59 @@ class NpzDataset(AtomicInMemoryDataset):
         return fields, fixed_fields
 
 
+def _ase_dataset_reader(
+    rank: int,
+    world_size: int,
+    tmpdir: str,
+    ase_kwargs: dict,
+    atomicdata_kwargs: dict,
+    include_frames,
+    global_options: dict,
+) -> Union[str, List[AtomicData]]:
+    """Parallel reader for all frames in file."""
+    if world_size > 1:
+        from nequip.utils._global_options import _set_global_options
+
+        # ^ avoid import loop
+        # we only `multiprocessing` if world_size > 1
+        _set_global_options(global_options)
+    # interleave--- in theory it is better for performance for the ranks
+    # to read consecutive blocks, but the way ASE is written the whole
+    # file gets streamed through all ranks anyway, so just trust the OS
+    # to cache things sanely, which it will.
+    # ASE handles correctly the case where there are no frames in index
+    # and just gives an empty list, so that will succeed:
+    index = slice(rank, None, world_size)
+    if include_frames is None:
+        # count includes 0, 1, ..., inf
+        include_frames = itertools.count()
+
+    datas = []
+    # stream them from ase too using iread
+    for i, atoms in enumerate(ase.io.iread(**ase_kwargs, index=index, parallel=False)):
+        global_index = rank + (world_size * i)
+        datas.append(
+            (
+                global_index,
+                AtomicData.from_ase(atoms=atoms, **atomicdata_kwargs)
+                if global_index in include_frames
+                # in-memory dataset will ignore this later, but needed for indexing to work out
+                else None,
+            )
+        )
+    # Save to a tempfile---
+    # there can be a _lot_ of tensors here, and rather than dealing with
+    # the complications of running out of file descriptors and setting
+    # sharing methods, since this is a one time thing, just make it simple
+    # and avoid shared memory entirely.
+    if world_size > 1:
+        path = f"{tmpdir}/rank{rank}.pth"
+        torch.save(datas, path)
+        return path
+    else:
+        return datas
+
+
 class ASEDataset(AtomicInMemoryDataset):
     """
 
@@ -790,10 +848,11 @@ class ASEDataset(AtomicInMemoryDataset):
         key_mapping: Optional[dict] = None,
         include_keys: Optional[List[str]] = None,
     ):
-
-        self.ase_args = dict(index=":")
+        self.ase_args = {}
         self.ase_args.update(getattr(type(self), "ASE_ARGS", dict()))
         self.ase_args.update(ase_args)
+        assert "index" not in self.ase_args
+        assert "filename" not in self.ase_args
 
         self.include_keys = include_keys
         self.key_mapping = key_mapping
@@ -851,15 +910,9 @@ class ASEDataset(AtomicInMemoryDataset):
     def raw_dir(self):
         return dirname(abspath(self.file_name))
 
-    def get_atoms(self):
-        from ase.io import read as aseread
-
-        return aseread(self.raw_dir + "/" + self.raw_file_names[0], **self.ase_args)
-
     def get_data(self):
-
-        # Get our data
-        atoms_list = self.get_atoms()
+        ase_args = {"filename": self.raw_dir + "/" + self.raw_file_names[0]}
+        ase_args.update(self.ase_args)
 
         # skip the None arguments
         kwargs = dict(
@@ -868,17 +921,35 @@ class ASEDataset(AtomicInMemoryDataset):
         )
         kwargs = {k: v for k, v in kwargs.items() if v is not None}
         kwargs.update(self.extra_fixed_fields)
+        n_proc = num_tasks()
+        with tempfile.TemporaryDirectory() as tmpdir:
+            from nequip.utils._global_options import _get_latest_global_options
 
-        if self.include_frames is None:
-            return (
-                [AtomicData.from_ase(atoms=atoms, **kwargs) for atoms in atoms_list],
+            # ^ avoid import loop
+            reader = functools.partial(
+                _ase_dataset_reader,
+                world_size=n_proc,
+                tmpdir=tmpdir,
+                ase_kwargs=ase_args,
+                atomicdata_kwargs=kwargs,
+                include_frames=self.include_frames,
+                # get the global options of the parent to initialize the worker correctly
+                global_options=_get_latest_global_options(),
             )
-        else:
-            return (
-                [
-                    AtomicData.from_ase(atoms=atoms_list[i], **kwargs)
-                    if i in self.include_frames
-                    else None  # in-memory dataset will ignore this later, but needed for indexing to work out
-                    for i in range(len(atoms_list))
-                ],
-            )
+            if n_proc > 1:
+                # things hang for some obscure OpenMP reason on some systems when using `fork` method
+                ctx = mp.get_context("forkserver")
+                with ctx.Pool(processes=n_proc) as p:
+                    # map it over the `rank` argument
+                    datas = p.map(reader, list(range(n_proc)))
+                # clean up the pool before loading the data
+                datas = [torch.load(d) for d in datas]
+                datas = sum(datas, [])
+                # un-interleave the datas
+                datas = sorted(datas, key=lambda e: e[0])
+            else:
+                datas = reader(rank=0)
+                # datas here is already in order, stride 1 start 0
+                # no need to un-interleave
+        # return list of AtomicData:
+        return ([e[1] for e in datas],)